Move users views into their own file.
