+++ zproject/urls.py
@@ -47,7 +47,7 @@ urlpatterns = patterns('',
         {'template_name': 'zerver/reset_done.html'}),
 
     # Avatar
-    url(r'^avatar/(?P<email>[\S]+)?', 'zerver.views.avatar'),
+    url(r'^avatar/(?P<email>[\S]+)?', 'zerver.views.users.avatar'),
 
     # Registration views, require a confirmation ID.
     url(r'^accounts/home/', 'zerver.views.accounts_home'),
@@ -176,9 +176,6 @@ urlpatterns += patterns('zerver.views',
 v1_api_and_json_patterns = patterns('zerver.views',
     url(r'^export$', 'rest_dispatch',
             {'GET':  'export'}),
-    url(r'^users$', 'rest_dispatch',
-            {'GET': 'get_members_backend',
-             'POST': 'create_user_backend'}),
     url(r'^users/me$', 'rest_dispatch',
             {'GET': 'get_profile_backend'}),
     url(r'^users/me/pointer$', 'rest_dispatch',
@@ -196,24 +193,29 @@ v1_api_and_json_patterns = patterns('zerver.views',
     url(r'^users/me/android_gcm_reg_id$', 'rest_dispatch',
         {'POST': 'add_android_reg_id',
          'DELETE': 'remove_android_reg_id'}),
-    url(r'^users/(?P<email>.*)/reactivate$', 'rest_dispatch',
-            {'POST': 'reactivate_user_backend'}),
-    url(r'^users/(?P<email>[^/]*)$', 'rest_dispatch',
-            {'PATCH': 'update_user_backend',
-             'DELETE': 'deactivate_user_backend'}),
-    url(r'^bots$', 'rest_dispatch',
-            {'GET': 'get_bots_backend',
-             'POST': 'add_bot_backend'}),
-    url(r'^bots/(?P<email>.*)/api_key/regenerate$', 'rest_dispatch',
-            {'POST': 'regenerate_bot_api_key'}),
-    url(r'^bots/(?P<email>.*)$', 'rest_dispatch',
-            {'PATCH': 'patch_bot_backend',
-             'DELETE': 'deactivate_bot_backend'}),
     url(r'^register$', 'rest_dispatch',
             {'POST': 'api_events_register'}),
 
     # Returns a 204, used by desktop app to verify connectivity status
     url(r'generate_204$', 'generate_204'),
+
+) + patterns('zerver.views.users',
+    url(r'^users$', 'rest_dispatch',
+        {'GET': 'get_members_backend',
+         'POST': 'create_user_backend'}),
+    url(r'^users/(?P<email>.*)/reactivate$', 'rest_dispatch',
+        {'POST': 'reactivate_user_backend'}),
+    url(r'^users/(?P<email>[^/]*)$', 'rest_dispatch',
+        {'PATCH': 'update_user_backend',
+         'DELETE': 'deactivate_user_backend'}),
+    url(r'^bots$', 'rest_dispatch',
+        {'GET': 'get_bots_backend',
+         'POST': 'add_bot_backend'}),
+    url(r'^bots/(?P<email>.*)/api_key/regenerate$', 'rest_dispatch',
+        {'POST': 'regenerate_bot_api_key'}),
+    url(r'^bots/(?P<email>.*)$', 'rest_dispatch',
+        {'PATCH': 'patch_bot_backend',
+         'DELETE': 'deactivate_bot_backend'}),
 
 ) + patterns('zerver.views.messages',
     # GET returns messages, possibly filtered, POST sends a message
